requirements.txt
torch
torchvision
Pillow
scikit-learn
tqdm
datasets
tensorflow
numpy